fix(vrf): migrate WitnessRoster to VRF v2.5; deploy core contracts
The roster was written against VRF v2 — uint64 subscription ids and positional
requestRandomWords args. v2.5 uses uint256 ids and a struct request with
ExtraArgsV1 encoding, so the first real request would have failed. A 252-bit
subscription id surfaced it.

DeployCore deploys everything except VerdictSink, whose CRE_REPORT_WRITER is
immutable and needs the Forwarder address confirmed first. It now requires
code.length > 0 before wiring: a partially broadcast deploy hit an RPC in-flight
limit, two contracts silently never deployed, and the one-time wireRegistry then
locked a roster to a codeless address permanently.

Deployed and wired on Sepolia; roster added as a VRF consumer, two agents
registered, one claim submitted with a bond. VRF fulfilment is outstanding — the
coordinator appears to be serving no requests at all.

+++ contracts/script/Deploy.s.sol
@@ -15,7 +15,7 @@ contract Deploy is Script {
         // VRF v2.5 Sepolia. Values from the Chainlink docs / your subscription.
         address vrfCoordinator = vm.envAddress("VRF_COORDINATOR");
         bytes32 keyHash = vm.envBytes32("VRF_KEY_HASH");
-        uint64 subId = uint64(vm.envUint("VRF_SUBSCRIPTION_ID"));
+        uint256 subId = vm.envUint("VRF_SUBSCRIPTION_ID");
         uint32 callbackGasLimit = uint32(vm.envOr("VRF_CALLBACK_GAS_LIMIT", uint256(500_000)));
 
         // ENSv2 hackathon deployment — NOT production ENS.
